Extracted meshRenderData to its own script
