+++ Sol/Source/GalaxyDraw/Interfaces/Renderer3D.cpp
@@ -1,7 +1,6 @@
 #include "solpch.h"
 #include "Renderer3D.h"
 
-#include "InstanceData.h"
 #include "Buffer.h"
 #include "VertexArray.h"
 #include "Shader.h"
@@ -11,7 +10,10 @@
 #include "RenderCommand.h"
 
 #include "Sol/Core/KeyedVector.h"
+
+#include "InstanceData.h"
 #include "TextureManager.h"
+#include "MeshRenderData.h"
 
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
@@ -20,24 +22,7 @@ namespace GalaxyDraw
 {
 
 
-	//for each unique mesh one of these structs are created,
-	//it contains the relevant buffers and data for rendering instances of said mesh.
-	struct MeshRenderData
-	{
-		std::string Name;
-		std::shared_ptr<Mesh> m_Mesh;
-		std::shared_ptr<VertexArray> m_VertexArray;
-		std::shared_ptr<VertexBuffer> m_VertexBuffer;
-		std::shared_ptr<IndexBuffer> m_IndexBuffer;
-		std::shared_ptr<InstanceBuffer> m_InstanceBuffer;
-
-		KeyedVector<EntityID, InstanceData> m_Instances;
-
-		Vertex* VertexBufferBase = nullptr;
-		Vertex* VertexBufferPtr = nullptr;
-		InstanceData* InstanceBufferBase = nullptr;
-		InstanceData* InstanceBufferPtr = nullptr;
-	};
+	
 
 
 	struct MaterialData
